combine duplicate getBestRealSplit logic
