+++ src/main/scala/io/citrine/lolo/trees/splits/ClassificationSplitter.scala
@@ -37,7 +37,7 @@ case class ClassificationSplitter(randomizedPivotLocation: Boolean = false, rng:
 
       /* Use different spliters for each type */
       val (possibleSplit, possibleImpurity) = rep._1(index) match {
-        case _: Double => getBestRealSplit(data, calculator, index, minInstances)
+        case _: Double => Splitter.getBestRealSplit[Char](data, calculator, index, minInstances, randomizedPivotLocation, rng)
         case _: Char => getBestCategoricalSplit(data, calculator, index, minInstances)
         case _: Any => throw new IllegalArgumentException("Trying to split unknown feature type")
       }
@@ -54,57 +54,6 @@ case class ClassificationSplitter(randomizedPivotLocation: Boolean = false, rng:
       val deltaImpurity = initialImpurity - bestImpurity
       (bestSplit, deltaImpurity)
     }
-  }
-
-  /**
-    * Find the best split on a continuous variable
-    *
-    * If randomizePivotLocation is true, the split pivots are drawn from a uniform random distribution between the two
-    * data points.  Each such pivot results in the same data split, but randomization can improve generalizability,
-    * particularly as part of an ensemble (i.e. random forests).
-    *
-    * @param data                 to split
-    * @param index                of the feature to split on
-    * @param minCount minimum number of data points to allow in each of the resulting splits
-    * @param randomizePivotLocation whether to generate splits randomly between the data points (default: false)
-    * @return the best split of this feature
-    */
-  def getBestRealSplit(
-                        data: Seq[(Vector[AnyVal], Char, Double)],
-                        calculator: GiniCalculator,
-                        index: Int,
-                        minCount: Int,
-                        randomizePivotLocation: Boolean = false
-                      ): (RealSplit, Double) = {
-    /* Pull out the feature that's considered here and sort by it */
-    val thinData = data.map(dat => (dat._1(index).asInstanceOf[Double], dat._2, dat._3)).sortBy(_._1)
-
-    var bestPurity = Double.MaxValue
-    var bestPivot = Double.MinValue
-
-    /* Move the data from the right to the left partition one value at a time */
-    calculator.reset()
-    (0 until data.size - minCount).foreach { j =>
-      val totalPurity = calculator.add(thinData(j)._2, thinData(j)._3)
-
-      /* Keep track of the best split, avoiding splits in the middle of constant sets of feature values
-         It is really important for performance to keep these checks together so
-         1) there is only one branch and
-         2) it is usually false
-       */
-      if (totalPurity < bestPurity && j + 1 >= minCount && Math.abs((thinData(j + 1)._1 - thinData(j)._1) / thinData(j)._1) > 1.0e-9) {
-        bestPurity = totalPurity
-        /* Try pivots at the midpoints between consecutive member values */
-        val left = thinData(j + 1)._1
-        val right = thinData(j)._1
-        bestPivot = if (randomizePivotLocation) {
-          (left - right) * rng.nextDouble() + right
-        } else {
-          (left + right) / 2.0
-        }
-      }
-    }
-    (new RealSplit(index, bestPivot), bestPurity)
   }
 
   def getBestCategoricalSplit(
