+++ src/test/scala/io/citrine/lolo/trees/splits/SplitterTest.scala
@@ -60,7 +60,7 @@ class SplitterTest {
 
     val calculator = new VarianceCalculator(0.0, 0.0, 8.0)
 
-    val (_, variance) = RegressionSplitter().getBestRealSplit(trainingData, calculator, 0, 1)
+    val (_, variance) = Splitter.getBestRealSplit(trainingData, calculator, 0, 1)
     assert(variance == Double.MaxValue, "didn't expect to find a valid split")
   }
